various css tweaks and fixes
Add around 1rem of padding on the body, so the card doesn't touch the edges on small screens.

On the body, change height to min-height - this way, the content will not get cut off if it grows beneath the viewport.

I would move the properties on .wrapper to the body. You don't need to declare nowrap, as that is the default value for flex-wrap.

Remove the width on the card. Instead, give it a max-width of around 20rem to prevent it from getting too wide on larger screens.

On the image, remove the width and height. Add display: block and max-width: 100% - the max-width prevents it from overflowing its container.